+++ java-manta-it/src/test/java/com/joyent/manta/client/MantaClientJobIT.java
@@ -131,7 +131,7 @@ public class MantaClientJobIT {
     }
 
 
-    @Test(dependsOnMethods = { "createJob" })
+    @Test(dependsOnMethods = {"createJob"})
     public void canAddAndGetInputsFromIterator() throws IOException {
         MantaJob job = buildJob();
         UUID jobId = mantaClient.createJob(job);